refactor: isolate pre-edit tool implementation
